[PATCH] pki/generate.ca: fix generated CA certs
Generate valid CA certs.  When the cert_signing_key is present clients
fail to validate the cert chain when connecting if trust=strict.
The error message indicates that the certificate violates the signer's
constraints.  Dropping this option generates certificates which
validate successfully.

While here adjust the CN to make it obvious that this particular cert
is intended to be used for signing other certs.

Signed-off-by: Kent R. Spillner <kspillner@acm.org>